[cloud-provider-dvp] move default SC discovery logic to module hook
