+++ modules/030-cloud-provider-dvp/hooks/discover.go
@@ -20,12 +20,14 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"log/slog"
 	"regexp"
 	"sort"
 	"strings"
 	"unicode"
 
 	"github.com/flant/addon-operator/pkg/module_manager/go_hook"
+	"github.com/flant/addon-operator/pkg/module_manager/go_hook/metrics"
 	"github.com/flant/addon-operator/sdk"
 	"github.com/flant/shell-operator/pkg/kube_events_manager/types"
 	corev1 "k8s.io/api/core/v1"
@@ -231,7 +233,65 @@ func getStorageClassName(value string) string {
 }
 
 func setStorageClassesValues(input *go_hook.HookInput, storageClasses []storageClass) {
+	const (
+		discoveryPath = "global.discovery.cloudProviderDefaultStorageClass"
+		metricName    = "d8_cloud_provider_dvp_default_storage_class_drifted"
+		metricGroup   = "cloud_provider_dvp_default_storage_class"
+	)
+
 	input.Values.Set("cloudProviderDvp.internal.storageClasses", storageClasses)
+
+	// Find and set default StorageClass in global values
+	var defaultSC string
+	for _, sc := range storageClasses {
+		if sc.IsDefault {
+			defaultSC = sc.Name
+			break
+		}
+	}
+
+	if defaultSC != "" {
+		input.Values.Set(discoveryPath, defaultSC)
+		input.Logger.Info("Discovered default storage class from DVP cloud provider", slog.String("storage_class", defaultSC))
+
+		// Detect drift
+		detectAndReportDrift(input, defaultSC, metricName, metricGroup)
+	} else {
+		input.Logger.Info("No default storage class found in parent DVP cluster")
+		input.Values.Remove(discoveryPath)
+		input.MetricsCollector.Expire(metricName)
+	}
+}
+
+// detectAndReportDrift compares expected default SC with actual and reports metric if drifted
+func detectAndReportDrift(input *go_hook.HookInput, expectedSC, metricName, metricGroup string) {
+	actualDefaultSC := input.Values.Get("global.discovery.defaultStorageClass").String()
+
+	// No actual default SC in cluster yet - no drift
+	if actualDefaultSC == "" {
+		input.MetricsCollector.Expire(metricName)
+		return
+	}
+
+	// Check for drift
+	if actualDefaultSC != expectedSC {
+		input.Logger.Warn("Default storage class drift detected",
+			slog.String("expected", expectedSC),
+			slog.String("actual", actualDefaultSC),
+		)
+		input.MetricsCollector.Set(
+			metricName,
+			1.0,
+			map[string]string{
+				"expected": expectedSC,
+				"actual":   actualDefaultSC,
+			},
+			metrics.WithGroup(metricGroup),
+		)
+	} else {
+		// No drift - expire the metric
+		input.MetricsCollector.Expire(metricName)
+	}
 }
 
 type storageClass struct {
